% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.R
\name{batch}
\alias{batch}
\title{Get the MIR and related information from the files}
\usage{
batch(file, mz1, mz2)
}
\arguments{
\item{file}{data file, CDF or other format supportted by xcmsRaw}

\item{mz1}{the lowest mass}

\item{mz2}{the highest mass}
}
\value{
Molecular isotope ratio
}
\description{
Get the MIR and related information from the files
}

