% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmzrt.R
\name{c.mzrt}
\alias{c.mzrt}
\title{c method for mzrt object}
\usage{
\method{c}{mzrt}(x, ...)
}
\arguments{
\item{x}{an mzrt object}

\item{...}{further mzrt objects to concatenate}
}
\value{
a combined mzrt object
}
\description{
c method for mzrt object
}
