% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{combinemd}
\alias{combinemd}
\title{Combine two or more matrix}
\usage{
combinemd(data1, data2, ...)
}
\arguments{
\item{data1}{matrix lower mass range}

\item{data2}{matrix higher mass range}

\item{...}{matrix even higher mass range}
}
\value{
matrix with the row as scantime in second and column as m/z
}
\description{
Combine two or more matrix
}

