% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{comparems}
\alias{comparems}
\title{substrict two GC-MS data}
\usage{
comparems(data1, data2, ...)
}
\arguments{
\item{data1}{the first data}

\item{data2}{the second data and will substrict the first one}
}
\description{
substrict two GC-MS data
}

