% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{dotpanno}
\alias{dotpanno}
\title{Perform MS/MS dot product annotation for mgf file}
\usage{
dotpanno(file, db = NULL, ppm = 10, prems = 1.1, binstep = 1, consinc = 0.6)
}
\arguments{
\item{file}{mgf file generated from MS/MS data}

\item{db}{database could be list object from `getMSP`}

\item{ppm}{mass accuracy, default 10}

\item{prems}{precursor mass range, default 1.1 to include M+H or M-H}

\item{binstep}{bin step for consin similarity}

\item{consinc}{consin similarity cutoff for annotation. Default 0.6.}
}
\value{
list with MSMS annotation results
}
\description{
Perform MS/MS dot product annotation for mgf file
}
