% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda.R
\name{findmet}
\alias{findmet}
\title{Screen metabolites by Mass Defect}
\usage{
findmet(list, mass, mdr = 50)
}
\arguments{
\item{list}{list with data as peaks list, mz, rt and group information, retention time should be in seconds}

\item{mass}{mass to charge ratio of specific compounds}

\item{mdr}{mass defect range, default 50mDa}
}
\value{
list with filtered metabolites mass to charge index of certain compound
}
\description{
Screen metabolites by Mass Defect
}
