% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fish}
\alias{fish}
\title{Fish experimental with technique and biological replicates}
\format{Large XCMSnExp with 996 peaks among 189 samples}
\usage{
fish
}
\description{
A XCMSnExp object containing 3 days' different treatment with 3 biological replicates, each biological replicate contain 6-8 technique replicates. Use as(fish,"xcmsSet") to convert them to xcmsSet object.
}
\keyword{datasets}
