% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fish2}
\alias{fish2}
\title{One factor fish experimental with technique and biological replicates}
\format{Large XCMSnExp with 992 peaks among 63 samples}
\usage{
fish2
}
\description{
A XCMSnExp object containing 1 days' different treatment with 3 biological replicates, each biological replicate contain 6-8 technique replicates. Use as(fish,"xcmsSet") to convert them to xcmsSet object.
}
\keyword{datasets}
