% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fish3}
\alias{fish3}
\title{Fish experimental with technique and biological replicates}
\format{Large XCMSnExp with 982 peaks among 20 samples}
\usage{
fish3
}
\description{
A XCMSnExp object containing 1 days' control samples with 3 biological replicates, each biological replicate contain 6-8 technique replicates. Use as(fish,"xcmsSet") to convert them to xcmsSet object.
}
\keyword{datasets}
