% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fish4}
\alias{fish4}
\title{Fish experimental with biological replicates}
\format{Large XCMSnExp with 408 peaks among 6 samples}
\usage{
fish4
}
\description{
A XCMSnExp object containing 1 days' control sample contain 6-8 technique replicates. Use as(fish,"xcmsSet") to convert them to xcmsSet object.
}
\keyword{datasets}
