% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sccp.R
\name{getarea}
\alias{getarea}
\title{Get the peak information from samples for SCCPs detection}
\usage{
getarea(data, ismz = 323, ppm = 5, rt = NULL, rts = NULL)
}
\arguments{
\item{data}{list from `xcmsRaw` function}

\item{ismz}{internal standards m/z}

\item{ppm}{resolution of mass spectrum}

\item{rt}{retention time range of sccps}

\item{rts}{retention time range of internal standards}
}
\value{
list with peak information
}
\description{
Get the peak information from samples for SCCPs detection
}
\seealso{
\code{\link{getareastd}},\code{\link{getsccp}}
}
