% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sccp.R
\name{getareastd}
\alias{getareastd}
\title{Get the peak information from SCCPs standards}
\usage{
getareastd(data = NULL, ismz = 323, ppm = 5, con = 2000,
  rt = NULL, rts = NULL)
}
\arguments{
\item{data}{list from `xcmsRaw` function}

\item{ismz}{internal standards m/z}

\item{ppm}{resolution of mass spectrum}

\item{con}{concentration of standards}

\item{rt}{retention time range of sccps}

\item{rts}{retention time range of internal standards}
}
\value{
list with peak information
}
\description{
Get the peak information from SCCPs standards
}
\seealso{
\code{\link{getarea}},\code{\link{getsccp}}
}
