% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcmsdoe.R
\name{getbiorep}
\alias{getbiorep}
\title{Get the report for biological replicates without technique replicates.}
\usage{
getbiorep(xset, method = "medret", intensity = "into", file = NULL,
  rsdcf = 30, inscf = 5)
}
\arguments{
\item{xset}{the xcmsset object which for bio replicated sample in different groups}

\item{method}{parameter for groupval function}

\item{intensity}{parameter for groupval function}

\item{file}{file name for further annotation, default NULL}

\item{rsdcf}{rsd cutoff for peaks, default 30}

\item{inscf}{Log intensity cutoff for peaks, default 5}
}
\value{
dataframe with mean, standard deviation and RSD for biological replicates combined with raw data
}
\description{
Get the report for biological replicates without technique replicates.
}
