% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcmsplus.R
\name{getbiotechrep}
\alias{getbiotechrep}
\title{Get the report for technique replicates.}
\usage{
getbiotechrep(xset, method = "medret", intensity = "into")
}
\arguments{
\item{xset}{the xcmsset object which for all of your technique replicates for bio replicated sample}

\item{method}{parameter for groupval function}

\item{intensity}{parameter for groupval function}
}
\value{
dataframe with mean, standard deviation and RSD for those technique replicates & biological replicates combined with raw data
}
\description{
Get the report for technique replicates.
}
