% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmzrt.R
\name{getcompare}
\alias{getcompare}
\title{Align multiple peaks list to one peak list}
\usage{
getcompare(..., index = 1, ppm = 5, deltart = 5)
}
\arguments{
\item{...}{peaks list, mzrt objects}

\item{index}{numeric, the index of reference peaks.}

\item{ppm}{pmd mass accuracy, default 5}

\item{deltart}{retention time shift table, default 10 seconds}
}
\value{
list object with aligned mzrt objects
}
\description{
Align multiple peaks list to one peak list
}
