% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massdefect.R
\name{getcorstd}
\alias{getcorstd}
\title{Paired correlationship among peak list based on cluster analysis}
\usage{
getcorstd(list, rtcutoff = 9, submass = c(15.9949, 14.003074, 26.01568,
  14.01565, 43.00581, 30.01056, 34.96885, 78.91834), mdcutoff = 0.02)
}
\arguments{
\item{list}{a list with mzrt profile}

\item{rtcutoff}{cutoff of the distances in cluster}

\item{submass}{mass vector of sub structure of homologous series}

\item{mdcutoff}{mass defect cluster cutoff}
}
\value{
list with retention time cluster, std mass defect analysis dataframe based on max average correlation
}
\description{
Paired correlationship among peak list based on cluster analysis
}
\seealso{
\code{\link{getmassdefect}},\code{\link{plotkms}},\code{\link{getpaired}}
}
