% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzrt.R
\name{getdoe}
\alias{getdoe}
\title{Filter the data based on DoE, rsd, intensity and NA}
\usage{
getdoe(list, inscf = 5, rsdcf = 30, NAtrim = "1", tr = F)
}
\arguments{
\item{list}{list with data as mzrt profile, mz, rt and group information}

\item{inscf}{Log intensity cutoff for peaks, default 5}

\item{rsdcf}{the rsd cutoff of all peaks}

\item{NAtrim}{FALSE means no NA trim, TRUE means remove the peaks with NA, '1' means impute the NA peaks with 1, '0' means impute the NA with 0.}

\item{tr}{logical. TRUE means dataset with technical replicates at the base level folder}
}
\value{
dataframe with filtered peaks
}
\description{
Filter the data based on DoE, rsd, intensity and NA
}
