% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmzrt.R
\name{getdwtus}
\alias{getdwtus}
\title{Density weighted intensity for one sample}
\usage{
getdwtus(peak, n = 512, log = FALSE)
}
\arguments{
\item{peak}{peaks intensity one sample}

\item{n}{the number of equally spaced points at which the density is to be estimated, default 512}

\item{log}{log transformation}
}
\value{
Density weighted intensity for one sample
}
\description{
Density weighted intensity for one sample
}
\examples{
data(list)
getdwtus(list$data[,1])
}
