% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzrt.R
\name{getfeaturest}
\alias{getfeaturest}
\title{Get the features from t test, with p value, q value, rsd and power restriction}
\usage{
getfeaturest(list, power = 0.8, pt = 0.05, qt = 0.05, n = 3,
  inscf = 5, rsdcf = 30)
}
\arguments{
\item{list}{list with data as mzrt profile, mz, rt and group information (two groups)}

\item{power}{defined power}

\item{pt}{p value threshold}

\item{qt}{q value threshold, BH adjust}

\item{n}{sample numbers in one group}

\item{inscf}{Log intensity cutoff for peaks, default 5}

\item{rsdcf}{the rsd cutoff of all peaks}
}
\value{
dataframe with peaks fit the setting above
}
\description{
Get the features from t test, with p value, q value, rsd and power restriction
}
\examples{
\dontrun{
library(faahKO)
cdfpath <- system.file("cdf", package = "faahKO")
xset <- getdata(cdfpath, pmethod = ' ')
list <- getmzrt(xset)
getfeaturest(list)
}
}
