% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmzrt.R
\name{getfilter}
\alias{getfilter}
\title{Filter the data based on row and column index}
\usage{
getfilter(list, rowindex = NULL, colindex = NULL)
}
\arguments{
\item{list}{list with data as peaks list, mz, rt and group information}

\item{rowindex}{logical, row index to keep}

\item{colindex}{logical, column index to keep}
}
\value{
list with remain peaks, and filtered peaks index
}
\description{
Filter the data based on row and column index
}
\examples{
data(list)
li <- getdoe(list)
lif <- getfilter(li,rowindex = li$rsdindex)
}
\seealso{
\code{\link{getdata2}},\code{\link{getdata}}, \code{\link{getmzrt}}, \code{\link{getimputation}}, \code{\link{getmr}}
}
