% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda.R
\name{getmass}
\alias{getmass}
\title{Get the exact mass of the isotopologues from a chemical formula or reaction's isotope patterns with the highest abundances}
\usage{
getmass(data)
}
\arguments{
\item{data}{a chemical formula or reaction e.g. 'Cl-H', 'C2H4'}
}
\value{
numerical vector
}
\description{
Get the exact mass of the isotopologues from a chemical formula or reaction's isotope patterns with the highest abundances
}
\examples{
getmass('CH2')
}
