% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massdefect.R
\name{getmassdefect}
\alias{getmassdefect}
\title{Get mass defect with certain scaled factor}
\usage{
getmassdefect(mass, sf)
}
\arguments{
\item{mass}{vector of mass}

\item{sf}{scaled factors}
}
\value{
dataframe with mass, scaled mass and scaled mass defect
}
\description{
Get mass defect with certain scaled factor
}
