% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massdefect.R
\name{getmassdiff}
\alias{getmassdiff}
\title{Isotope extraction for single group of samples with certain mass diff}
\usage{
getmassdiff(list, massdiff, rtwindow, mzwindow, ppm)
}
\arguments{
\item{list}{a list with mzrt profile, mz and rt}

\item{massdiff}{mass defect}

\item{rtwindow}{retention time range}

\item{mzwindow}{mass charge ratio window}

\item{ppm}{resolution of the mass spectrum}
}
\value{
dataframe with mass, retention time, scaled mass and scaled mass defect
}
\description{
Isotope extraction for single group of samples with certain mass diff
}
\seealso{
\code{\link{getmassdefect}},\code{\link{plotkms}}
}
