% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda.R
\name{getmdh}
\alias{getmdh}
\title{Get the high order unit based Mass Defect}
\usage{
getmdh(mz, cus = c("CH2,H2"), method = "round")
}
\arguments{
\item{mz}{numeric vector for exact mass}

\item{cus}{chemical formula or reaction}

\item{method}{you could use `round`, `floor` or `ceiling`}
}
\value{
high order Mass Defect with details
}
\description{
Get the high order unit based Mass Defect
}
\examples{
\dontrun{
getmdh(getmass('C2H4')))
}
}
