% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda.R
\name{getmdr}
\alias{getmdr}
\title{Get the raw Mass Defect}
\usage{
getmdr(mz)
}
\arguments{
\item{mz}{numeric vector for exact mass}
}
\value{
raw Mass Defect
}
\description{
Get the raw Mass Defect
}
\examples{
getmdr(getmass('C2H4'))
}
