% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno.R
\name{getms1anno}
\alias{getms1anno}
\title{Annotation of MS1 data by compounds database by predefined paired mass distance}
\usage{
getms1anno(pmd, mz, ppm = 10, db = NULL)
}
\arguments{
\item{pmd}{adducts formula or paired mass distance for ions}

\item{mz}{unknown mass to charge ratios vector}

\item{ppm}{mass accuracy}

\item{db}{compounds database as dataframe. Two required columns are name and monoisotopic molecular weight with column names of name and mass}
}
\value{
list or data frame
}
\description{
Annotation of MS1 data by compounds database by predefined paired mass distance
}
