% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{getMSP}
\alias{getMSP}
\title{read in MSP file as list for ms/ms or ms(EI) annotation}
\usage{
getMSP(file)
}
\arguments{
\item{file}{the path to your MSP file}
}
\value{
list a list with MSP information for annotation
}
\description{
read in MSP file as list for ms/ms or ms(EI) annotation
}
