% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{getmzrt}
\alias{getmzrt}
\title{Get the mzrt profile and group information for batch correction and plot as a list}
\usage{
getmzrt(xset)
}
\arguments{
\item{xset}{xcmsSet objects}
}
\value{
list with rtmz profile and group infomation
}
\description{
Get the mzrt profile and group information for batch correction and plot as a list
}
\examples{
\dontrun{
library(faahKO)
cdfpath <- system.file("cdf", package = "faahKO")
xset <- getdata(cdfpath, pmethod = ' ')
getmzrt(xset)
}
}
