% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmzrt.R
\name{getoverlapmass}
\alias{getoverlapmass}
\title{Get the overlap peaks by mass range}
\usage{
getoverlapmass(mzrange1, mzrange2)
}
\arguments{
\item{mzrange1}{mass range 1 to be overlapped}

\item{mzrange2}{mass range 2 to overlap}
}
\value{
logical index for mzrange1's peaks
}
\description{
Get the overlap peaks by mass range
}
\seealso{
\code{\link{getmzrt}}, \code{\link{getimputation}}, \code{\link{getmr}},\code{\link{getdoe}}, \code{\link{getoverlappeak}},\code{\link{getoverlaprt}}
}
