% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno.R
\name{getoverlappeak}
\alias{getoverlappeak}
\title{Get the overlap peaks by mass and retention time range}
\usage{
getoverlappeak(list1, list2)
}
\arguments{
\item{list1}{list with data as peaks list, mz, rt, mzrange, rtrange and group information to be overlapped}

\item{list2}{list with data as peaks list, mz, rt, mzrange, rtrange and group information to overlap}
}
\value{
logical index for list 1's peaks
}
\description{
Get the overlap peaks by mass and retention time range
}
\seealso{
\code{\link{getimputation}},\code{\link{getdoe}}
}
