% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmzrt.R
\name{getpn}
\alias{getpn}
\title{Merge positive and negative mode data}
\usage{
getpn(pos, neg, ppm = 5, pmd = 2.02, digits = 2, cutoff = 0.9)
}
\arguments{
\item{pos}{a list with mzrt profile collected from positive mode. The sample order should match the negative mode.}

\item{neg}{a list with mzrt profile collected from negative mode.The sample order should match the positive mode.}

\item{ppm}{pmd mass accuracy, default 5}

\item{pmd}{numeric or numeric vector}

\item{digits}{mass or mass to charge ratio accuracy for pmd, default 2}

\item{cutoff}{correlation coefficients, default 0.9}
}
\value{
mzrt object with group information from pos mode
}
\description{
Merge positive and negative mode data
}
