% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmzrt.R
\name{getpqsi}
\alias{getpqsi}
\title{Compute pooled QC linear index according to run order}
\usage{
getpqsi(data, order, n = 5)
}
\arguments{
\item{data}{peaks intensity list with row as peaks and column as samples}

\item{order}{run order of pooled QC samples}

\item{n}{samples numbers used for linear regression}
}
\value{
vector for the peaks proportion with significant changes in linear regression after FDR control.
}
\description{
Compute pooled QC linear index according to run order
}
