% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcmsplus.R
\name{getQCraw}
\alias{getQCraw}
\title{get the data of QC compound for a group of data}
\usage{
getQCraw(path, mzrange, rtrange, index = NULL)
}
\arguments{
\item{path}{data path for your QC samples}

\item{mzrange}{mass of the QC compound}

\item{rtrange}{retention time of the QC compound}

\item{index}{index of the files contained QC compounds, default is all of the compounds}
}
\value{
number vector, each number indicate the peak area of that mass and retention time range
}
\description{
get the data of QC compound for a group of data
}
