% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmzrt.R
\name{getrangecsv}
\alias{getrangecsv}
\title{Get a mzrt list and/or save mz and rt range as csv file}
\usage{
getrangecsv(list, name, ...)
}
\arguments{
\item{list}{list with data as peaks list, mz, rt and group information}

\item{name}{result name for csv and/or eic file, default NULL}

\item{...}{other parameters for `write.table`}
}
\value{
NULL, csv file
}
\description{
Get a mzrt list and/or save mz and rt range as csv file
}
