% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda.R
\name{getrmd}
\alias{getrmd}
\title{Get the Relative Mass Defect}
\usage{
getrmd(mz)
}
\arguments{
\item{mz}{numeric vector for exact mass}
}
\value{
Relative Mass Defect
}
\description{
Get the Relative Mass Defect
}
\examples{
getrmd(getmass('C2H4'))
}
