% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalstd.R
\name{getsda}
\alias{getsda}
\title{Perform structure directed analysis for peaks list.}
\usage{
getsda(list, rtcutoff = 3, freqcutoff = 10, pmdcutoff = 500)
}
\arguments{
\item{list}{a list with mzrt profile}

\item{rtcutoff}{cutoff of the retention time in secounds, default 3}

\item{freqcutoff}{cutoff of the paired mass difference frequency, default 10}

\item{pmdcutoff}{cutoff of the largest paired mass differences, default 500}
}
\value{
list with tentative isotope, adducts, and neutral loss peaks' index, retention time clusters.
}
\description{
Perform structure directed analysis for peaks list.
}
\seealso{
\code{\link{getpaired}},\code{\link{getstd}},\code{\link{plotpaired}}
}
