% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcmsplus.R
\name{getsim}
\alias{getsim}
\title{output the similarity of two dataset}
\usage{
getsim(xset1, xset2)
}
\arguments{
\item{xset1}{the first dataset}

\item{xset2}{the second dateset}
}
\value{
similarity on retention time and rsd %
}
\description{
output the similarity of two dataset
}
