% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massdefect.R
\name{getstd}
\alias{getstd}
\title{Find the standard mass for each retention time hierarchical clustering based on PMD relationship within each retention time cluster and isotope and return the index of the std data for each retention time cluster.}
\usage{
getstd(list)
}
\arguments{
\item{list}{a list from getpaired function}
}
\value{
list with std mass index
}
\description{
Find the standard mass for each retention time hierarchical clustering based on PMD relationship within each retention time cluster and isotope and return the index of the std data for each retention time cluster.
}
\seealso{
\code{\link{getpaired}},\code{\link{getmdg}}
}
