% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{getsubmd}
\alias{getsubmd}
\title{Subset the data mass spectrum of certain retention time and plot them}
\usage{
getsubmd(data, rt, ms)
}
\arguments{
\item{data}{imported data matrix of GC-MS}

\item{rt}{vector range of the retention time min}

\item{ms}{vector range of the m/z}
}
\value{
data matrix
}
\description{
Subset the data mass spectrum of certain retention time and plot them
}

