% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{gettechbiorep}
\alias{gettechbiorep}
\title{Get the report for samples with technique replicates}
\usage{
gettechbiorep(xset, anno = F, peaklist = F, file = NULL,
  method = "medret", intensity = "into")
}
\arguments{
\item{xset}{the xcmsset object all of samples with technique replicates}

\item{anno}{logical if set as True, it will return the table for further annotation, default false}

\item{peaklist}{logical if set as True, it will return csv files for metaboanalyst, default false}

\item{file}{file name for the peaklist}

\item{method}{parameter for groupval function}

\item{intensity}{parameter for groupval function}
}
\value{
dataframe with mean, standard deviation and RSD for those technique replicates combined with raw data for all of the samples if anno and peaklist are defaults false.
}
\description{
Get the report for samples with technique replicates
}
