% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcmsdoe.R
\name{gettimegrouprep}
\alias{gettimegrouprep}
\title{Get the time series or two factor DoE report for samples with biological and technique replicates in different groups}
\usage{
gettimegrouprep(xset, file = NULL, method = "medret", intensity = "into",
  rsdcf = 30, inscf = 5)
}
\arguments{
\item{xset}{the xcmsset object all of samples with technique replicates in time series or two factor DoE}

\item{file}{file name for the peaklist to MetaboAnalyst}

\item{method}{parameter for groupval function}

\item{intensity}{parameter for groupval function}

\item{rsdcf}{rsd cutoff for peaks, default 30}

\item{inscf}{Log intensity cutoff for peaks, default 5}
}
\value{
dataframe with time series or two factor DoE mean, standard deviation and RSD for those technique replicates & biological replicates combined with raw data in different groups if file are defaults NULL.
}
\description{
Get the time series or two factor DoE report for samples with biological and technique replicates in different groups
}
