% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{getupload2}
\alias{getupload2}
\title{Get the csv files to be submitted to Metaboanalyst}
\usage{
getupload2(xset, value = "into", name = "Peaklist")
}
\arguments{
\item{xset}{a XCMSnExp object with processed data which you want to submitted to Metaboanalyst}

\item{value}{value for `xcms::featureValues`}

\item{name}{file name}
}
\value{
dataframe with data needed for Metaboanalyst if your want to perform local analysis.
}
\description{
Get the csv files to be submitted to Metaboanalyst
}
\examples{
\dontrun{
library(faahKO)
cdfpath <- system.file('cdf', package = 'faahKO')
xset <- getdata2(cdfpath,
ppp = xcms::MatchedFilterParam(),
rtp = xcms::ObiwarpParam(),
gpp = xcms::PeakDensityParam())
getupload2(xset)
}
}
\seealso{
\code{\link{getdata2}},\code{\link{getupload}}, \code{\link{getmzrt2}}
}
