% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{getupload3}
\alias{getupload3}
\title{Get the csv files to be submitted to Metaboanalyst}
\usage{
getupload3(list, name = "Peaklist")
}
\arguments{
\item{list}{list with data as peaks list, mz, rt and group information}

\item{name}{file name}
}
\value{
dataframe with data needed for Metaboanalyst if your want to perform local analysis.
}
\description{
Get the csv files to be submitted to Metaboanalyst
}
\examples{
\dontrun{
library(faahKO)
cdfpath <- system.file('cdf', package = 'faahKO')
xset <- getdata2(cdfpath,
ppp = xcms::MatchedFilterParam(),
rtp = xcms::ObiwarpParam(),
gpp = xcms::PeakDensityParam())
xset <- enviGCMS::getmzrt2(xset)
getupload3(xset)
}
}
\seealso{
\code{\link{getmzrt}}, \code{\link{getmzrt2}}
}
