% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmzrt.R
\name{gifmr}
\alias{gifmr}
\title{plot scatter plot for rt-mz profile and output gif file for multiple groups}
\usage{
gifmr(
  list,
  ms = c(100, 500),
  rsdcf = 30,
  inscf = 5,
  imputation = "l",
  name = "test",
  ...
)
}
\arguments{
\item{list}{list with data as peaks list, mz, rt and group information}

\item{ms}{the mass range to plot the data}

\item{rsdcf}{the rsd cutoff of all peaks in all group}

\item{inscf}{Log intensity cutoff for peaks across samples. If any peaks show a intensity higher than the cutoff in any samples, this peaks would not be filtered. default 5}

\item{imputation}{parameters for `getimputation` function method}

\item{name}{file name for gif file, default test}

\item{...}{parameters for `plot` function}
}
\value{
gif file
}
\description{
plot scatter plot for rt-mz profile and output gif file for multiple groups
}
\examples{
\dontrun{
data(list)
gifmr(list)
}
}
