% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massdefect.R
\name{globalstd}
\alias{globalstd}
\title{GlobalStd algorithm}
\usage{
globalstd(list, rtcutoff = 9, isocutoff = 3, freqcutoff = 20,
  corcutoff = 0.8, submass = c(15.9949, 14.003074, 26.01568, 14.01565,
  43.00581, 30.01056, 34.96885, 78.91834), mdcutoff = 0.02, mdgn = 50,
  lv = NULL)
}
\arguments{
\item{list}{a peaks list with mass to charge}

\item{rtcutoff}{cutoff of the distances in cluster}

\item{isocutoff}{cutoff to find the isotope relationship}

\item{freqcutoff}{cutoff of the mass differences frequency}

\item{corcutoff}{cutoff of the correlation coefficient, default 0.8}

\item{submass}{mass vector of sub structure of homologous series}

\item{mdcutoff}{mass defect cluster cutoff}

\item{mdgn}{mass defect groups numbers for interval, 50 means 0.02 inteval on mass defect scale from -0.5 to 0.5}

\item{lv}{group info for the data}
}
\value{
list with GlobalStd algorithm processed data.
}
\description{
GlobalStd algorithm
}
\seealso{
\code{\link{getpaired}},\code{\link{getstd}},\code{\link{getmdg}}
}
