% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{integration}
\alias{integration}
\title{Just integrate data according to fixed rt and fixed noise area}
\usage{
integration(data, rt = c(8.3, 9), brt = c(8.3, 8.4), smoothit = TRUE)
}
\arguments{
\item{data}{file should be a dataframe with the first column RT and second column intensity of the SIM ions.}

\item{rt}{a rough RT range contained only one peak to get the area}

\item{brt}{a rough RT range contained only one peak and enough noises to get the area}

\item{smoothit}{logical, if using an average smooth box or not. If using, n will be used}
}
\value{
area integration data
}
\description{
Just integrate data according to fixed rt and fixed noise area
}
\examples{
\dontrun{
area <- integration(data)
}
}
