% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\docType{data}
\name{list}
\alias{list}
\title{Demo data}
\format{
A list object with data, mass to charge ratio, retention time and group information. The list is generated from faahKO package.
}
\usage{
data(list)
}
\description{
Demo data
}
\keyword{datasets}
