% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/code.R
\name{ma}
\alias{ma}
\title{filter data by average moving box}
\usage{
ma(x, n)
}
\arguments{
\item{x}{a vector}

\item{n}{A number to indentify the size of the moving box.}
}
\value{
The filtered data
}
\description{
filter data by average moving box
}
\examples{
ma(c(1:10), 2)
}

