% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\docType{data}
\name{matrix}
\alias{matrix}
\title{Demo raw data matrix}
\format{
A matrix object from raw mass spectrometry data. The list is generated from faahKO package.
}
\usage{
data(matrix)
}
\description{
Demo raw data matrix
}
\keyword{datasets}
