% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsanno.R
\name{plotanno}
\alias{plotanno}
\title{Show MS/MS pmd annotation result}
\usage{
plotanno(anno, ...)
}
\arguments{
\item{anno}{list from MSmS anno function}

\item{...}{other parameter for plot function}
}
\description{
Show MS/MS pmd annotation result
}
