% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{plotc}
\alias{plotc}
\title{plot the comparision of two xcmsset according to certain method}
\usage{
plotc(data1, data2, method = "max", p = 0.5, mzrange = c(100, 1000))
}
\arguments{
\item{data1}{the first xcmsset data set}

\item{data2}{the second xcmsset data set}

\item{method}{could be 'max', 'mean', 'median'}

\item{p}{filtered peaks larger than the propotion of the max, mean or median}

\item{mzrange}{the mass range to plot the data}
}
\description{
plot the comparision of two xcmsset according to certain method
}
