% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{plotcc}
\alias{plotcc}
\title{Plot the calibration curve with error bar, r squared and equation}
\usage{
plotcc(x, y, upper, lower = upper, ...)
}
\arguments{
\item{x}{concentration}

\item{y}{response}

\item{upper}{upper error bar}

\item{lower}{lower error bar}

\item{...}{parameters for `plot` function}
}
\description{
Plot the calibration curve with error bar, r squared and equation
}
\examples{
\dontrun{
plotcc(x,y,upper)
}
}
