% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmzrt.R
\name{plotdwtus}
\alias{plotdwtus}
\title{plot density weighted intensity for multiple samples}
\usage{
plotdwtus(list, n = 512, ...)
}
\arguments{
\item{list}{list with data as peaks list, mz, rt and group information}

\item{n}{the number of equally spaced points at which the density is to be estimated, default 512}

\item{...}{parameters for `plot` function}
}
\value{
Density weighted intensity for multiple samples
}
\description{
plot density weighted intensity for multiple samples
}
\examples{
data(list)
plotdwtus(list)
}
