% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{plote}
\alias{plote}
\title{plot EIC and boxplot for all peaks and return diffreport}
\usage{
plote(xset, name = "test", test = "t", nonpara = "n", ...)
}
\arguments{
\item{xset}{xcmsset object}

\item{name}{filebase of the sub dir}

\item{test}{'t' means two-sample welch t-test, 't.equalvar' means two-sample welch t-test with equal variance, 'wilcoxon' means rank sum wilcoxon test, 'f' means F-test, 'pairt' means paired t test, 'blockf' means Two-way analysis of variance, default 't'}

\item{nonpara}{'y' means using nonparametric ranked data, 'n' means original data}

\item{...}{other parameters for `diffreport`}
}
\value{
diffreport and pdf figure for EIC and boxplot
}
\description{
plot EIC and boxplot for all peaks and return diffreport
}
\examples{
\dontrun{
library(faahKO)
cdfpath <- system.file('cdf', package = 'faahKO')
xset <- getdata(cdfpath, pmethod = ' ')
plote(xset)
}
}
