% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{plotgroup}
\alias{plotgroup}
\title{Plot the response group of GC-MS}
\usage{
plotgroup(data, threshold = 2)
}
\arguments{
\item{data}{imported data matrix of GC-MS}

\item{threshold}{the threshold of the response (log based 10) to seperate the group}
}
\value{
list linear regression model for the data matrix
}
\description{
Plot the response group of GC-MS
}
\examples{
\dontrun{
data <- getmd(rawdata)
plotgroup(data)
}
}
