% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{plothist}
\alias{plothist}
\title{Plot the density of the GC-MS data with EM algorithm to separate the data into two log normal distribution}
\usage{
plothist(data)
}
\arguments{
\item{data}{imported data matrix of GC-MS}
}
\description{
Plot the density of the GC-MS data with EM algorithm to separate the data into two log normal distribution
}
\examples{
\dontrun{
# generate a matrix from raw data with row as m/z and column as retention time
plothist(matrix)
}
}
