% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmzrt.R
\name{plothm}
\alias{plothm}
\title{Plot the heatmap of mzrt profiles}
\usage{
plothm(data, lv, index = NULL)
}
\arguments{
\item{data}{data row as peaks and column as samples}

\item{lv}{group information}

\item{index}{index for selected peaks}
}
\description{
Plot the heatmap of mzrt profiles
}
\examples{
data(list)
plothm(list$data, lv = as.factor(list$group$sample_group))
}
