% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{plotint}
\alias{plotint}
\title{plot the information of integration}
\usage{
plotint(list, name = NULL)
}
\arguments{
\item{list}{list from getinteagtion}

\item{name}{the title of the plot}
}
\description{
plot the information of integration
}
\examples{
\dontrun{
list <- getinteagtion(rawdata)
plotint(list)
}
}
