% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{plotintslope}
\alias{plotintslope}
\title{plot the slope information of intergretion}
\usage{
plotintslope(list, name = NULL)
}
\arguments{
\item{list}{list from getinteragtion}

\item{name}{the title of the plot}
}
\description{
plot the slope information of intergretion
}
\examples{
\dontrun{
list <- getinteragtion(rawdata)
plotintslope(list)
}
}
