% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{plotkms}
\alias{plotkms}
\title{plot the kendrick mass defect diagram}
\usage{
plotkms(data, cutoff = 1000)
}
\arguments{
\item{data}{vector with the name m/z}

\item{cutoff}{remove the low intensity}
}
\description{
plot the kendrick mass defect diagram
}

