% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzrt.R
\name{plotmrc}
\alias{plotmrc}
\title{plot the diff scatter plot for one xcmsset objects with threshold and two groups}
\usage{
plotmrc(xset, ms = c(100, 800), inscf = 5, rsdcf = 30, ...)
}
\arguments{
\item{xset}{xcmsset object with two groups}

\item{ms}{the mass range to plot the data}

\item{inscf}{Log intensity cutoff for peaks, default 5}

\item{rsdcf}{the rsd cutoff of all peaks}

\item{...}{parameters for `plot` function}
}
\description{
plot the diff scatter plot for one xcmsset objects with threshold and two groups
}
\examples{
\dontrun{
library(faahKO)
cdfpath <- system.file("cdf", package = "faahKO")
xset <- getdata(cdfpath, pmethod = ' ')
plotmrc(xset)
}
}
