% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{plotms}
\alias{plotms}
\title{plot GC-MS data as a heatmap}
\usage{
plotms(data, col = heat.colors(108), ...)
}
\arguments{
\item{data}{imported data matrix of GC-MS}

\item{col}{custumized color}
}
\value{
heatmap
}
\description{
plot GC-MS data as a heatmap
}

