% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{plotmsrt}
\alias{plotmsrt}
\title{Plot EIC of certain m/z and return dataframe for intergration}
\usage{
plotmsrt(data, ms, rt = c(3.1, 25), n = F)
}
\arguments{
\item{data}{imported data matrix of GC-MS}

\item{ms}{m/z to be extracted}

\item{rt}{vector range of the retention time}

\item{n}{logical smooth or not}
}
\value{
dataframe with  with the first column RT and second column intensity of the SIM ions.
}
\description{
Plot EIC of certain m/z and return dataframe for intergration
}

