% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalstd.R
\name{plotpaired}
\alias{plotpaired}
\title{Plot the mass pairs and high frequency mass differences}
\usage{
plotpaired(list, index = NULL, ...)
}
\arguments{
\item{list}{a list from getpaired function}

\item{index}{index for PMD value}

\item{...}{other parameters for plot function}
}
\description{
Plot the mass pairs and high frequency mass differences
}
\seealso{
\code{\link{getpaired}}, \code{\link{globalstd}}
}
