% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmzrt.R
\name{plotpeak}
\alias{plotpeak}
\title{plot intensity of peaks across samples or samples across peaks}
\usage{
plotpeak(data, lv = NULL, indexx = NULL, indexy = NULL, ...)
}
\arguments{
\item{data}{matrix}

\item{lv}{factor vector for the column}

\item{indexx}{index for matrix row}

\item{indexy}{index for matrix column}

\item{...}{parameters for `title` function}
}
\value{
parallel coordinates plot
}
\description{
plot intensity of peaks across samples or samples across peaks
}
\examples{
data(list)
# selected peaks across samples
plotpeak(t(list$data), lv = as.factor(c(rep(1,5),rep(2,nrow(list$data)-5))),1:10,1:10)
# selected samples across peaks
plotpeak(list$data, lv = as.factor(list$group$sample_group),1:10,1:10)
}
