% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmzrt.R
\name{plotridge}
\alias{plotridge}
\title{plot ridgeline density plot}
\usage{
plotridge(data, xlab = "", ylab = "", index = NULL)
}
\arguments{
\item{data}{matrix}

\item{xlab}{label for x axis}

\item{ylab}{label for y axis}

\item{index}{index for y}
}
\value{
ridgeline density plot
}
\description{
plot ridgeline density plot
}
\examples{
data(list)
plotridge(t(list$data),'Intensity','peaks',c(1:10))
}
