% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmzrt.R
\name{plotrla}
\alias{plotrla}
\title{Relative Log Abundance (RLA) plots}
\usage{
plotrla(data, lv, type = "g", ...)
}
\arguments{
\item{data}{data row as peaks and column as samples}

\item{lv}{factor vector for the group information}

\item{type}{'g' means group median based, other means all samples median based.}

\item{...}{parameters for boxplot}
}
\value{
Relative Log Abundance (RLA) plots
}
\description{
Relative Log Abundance (RLA) plots
}
\examples{
data(list)
plotrla(list$data, as.factor(list$group$sample_group))
}
