% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcmsplot.R
\name{plotrsd}
\alias{plotrsd}
\title{plot the rsd influnces of data}
\usage{
plotrsd(xset, ms = c(0, 800), type = "g", inscf = 5, rsdcf = 30, ...)
}
\arguments{
\item{xset}{xcmsset data}

\item{ms}{the mass range to plot the data}

\item{type}{'t' means biological samples with technique replicates; 'b' means biological samples without technique replicates in one group; 'g' means biological samples without technique replicates in mutiple groups. default 'g'}

\item{inscf}{Log intensity cutoff for peaks, default 5}

\item{rsdcf}{the rsd cutoff of all peaks}

\item{...}{other parameters for `plot` function}
}
\description{
plot the rsd influnces of data
}
\examples{
\dontrun{
library(faahKO)
cdfpath <- system.file("cdf", package = "faahKO")
xset <- getdata(cdfpath[1:6], pmethod = ' ')
plotrsd(xset)
}
}
