% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalstd.R
\name{plotrtg}
\alias{plotrtg}
\title{Plot the retention time group}
\usage{
plotrtg(list, ...)
}
\arguments{
\item{list}{a list from getpaired function}

\item{...}{other parameters for plot function}
}
\description{
Plot the retention time group
}
\seealso{
\code{\link{getpaired}}, \code{\link{globalstd}}
}
