% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{plotsms}
\alias{plotsms}
\title{Plot the intensity distribution of GC-MS}
\usage{
plotsms(meanmatrix, rsdmatrix)
}
\arguments{
\item{meanmatrix}{mean data matrix of GC-MS(n=5)}

\item{rsdmatrix}{standard deviation matrix of GC-MS(n=5)}
}
\description{
Plot the intensity distribution of GC-MS
}

