% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massdefect.R
\name{plotstdmd}
\alias{plotstdmd}
\title{Plot the std mass from GlobalStd algorithm in certain mass defect groups}
\usage{
plotstdmd(list, mdindex)
}
\arguments{
\item{list}{a list from getmdg function}

\item{mdindex}{mass defect index}
}
\description{
Plot the std mass from GlobalStd algorithm in certain mass defect groups
}
\seealso{
\code{\link{getstd}}, \code{\link{globalstd}},\code{\link{plotstd}},\code{\link{plotpaired}},\code{\link{plotstdrt}}
}
