% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{plotsub}
\alias{plotsub}
\title{Plot the background of data}
\usage{
plotsub(data)
}
\arguments{
\item{data}{imported data matrix of GC-MS}
}
\description{
Plot the background of data
}
\examples{
\dontrun{
plotsub(matrix)
}
}
