% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{plottic}
\alias{plottic}
\title{Plot Total Ion Chromatogram (TIC)}
\usage{
plottic(data, n = FALSE)
}
\arguments{
\item{data}{imported data matrix of GC-MS}

\item{n}{logical smooth or not}
}
\value{
plot
}
\description{
Plot Total Ion Chromatogram (TIC)
}
\examples{
\dontrun{
matrix <- getmd(rawdata)
plottic(matrix)
}
}
