% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{ridgesplot}
\alias{ridgesplot}
\title{Relative Log Abundance Ridge(RLAR) plots}
\usage{
ridgesplot(data, lv, type = "g")
}
\arguments{
\item{data}{data as mzrt profile}

\item{lv}{factor vector for the group infomation}

\item{type}{'g' means group median based, other means all samples median based.}
}
\value{
Relative Log Abundance (RLA) plots
}
\description{
Relative Log Abundance Ridge(RLAR) plots
}
\examples{
sim <- mzrtsim()
ridgesplot(sim$data, as.factor(sim$con))
}
