% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{rlaplot}
\alias{rlaplot}
\title{Relative Log Abundance (RLA) plots}
\usage{
rlaplot(data, lv, type = "g")
}
\arguments{
\item{data}{data as mzrt profile}

\item{lv}{factor vector for the group infomation}

\item{type}{'g' means group median based, other means all samples median based.}
}
\value{
Relative Log Abundance (RLA) plots
}
\description{
Relative Log Abundance (RLA) plots
}
\examples{
sim <- mzrtsim()
rlaplot(sim$data, as.factor(sim$con))
}
