% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sccp.R
\docType{data}
\name{sccp}
\alias{sccp}
\title{Short-Chain Chlorinated Paraffins(SCCPs) peaks information for quantitative analysis}
\format{
A data frame with 24 rows and 8 variables:
\describe{
  \item{Cln}{Chlorine atom numbers}
  \item{Cn}{Carbon atom numbers}
  \item{formula}{molecular formula}
  \item{Hn}{hydrogen atom numbers}
  \item{ions}{[M-Cl]- ions}
  \item{mz}{m/z for the isotopologues with highest intensity}
  \item{intensity}{abundance of the isotopologues with highest intensity}
  \item{Clp}{Chlorine contents}
}
}
\usage{
data(sccp)
}
\description{
A dataset containing the ions, formula, Cl%, peak abundances for all SCCPs compounds
}
\keyword{datasets}
