% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzrtsim.R
\name{simdata}
\alias{simdata}
\title{Save the simulated data as csv files}
\usage{
simdata(sim, name = "sim")
}
\arguments{
\item{sim}{list from `mzrtsim` or `simmzrt`}

\item{name}{file name}
}
\description{
Save the simulated data as csv files
}
\seealso{
\code{\link{mzrtsim}},\code{\link{simmzrt}}
}
