% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzrtsim.R
\name{simmzrt}
\alias{simmzrt}
\title{Simulation from data by sample mean and rsd from Empirical Cumulative Distribution or Bootstrap sampling}
\usage{
simmzrt(data, type = "e", npeaks = 1000, ncomp = 0.8, ncond = 2,
  ncpeaks = 0.05, nbatch = 3, nbpeaks = 0.1, npercond = 10,
  nperbatch = c(8, 5, 7), seed = 42)
}
\arguments{
\item{data}{matrix with row peaks and column samples}

\item{type}{'e' means simulation from data by sample mean and rsd from Empirical Cumulative Distribution; 'f' means simulation from data by sample mean and rsd from Bootstrap sampling}

\item{npeaks}{Number of peaks to simulate}

\item{ncomp}{percentage of compounds}

\item{ncond}{Number of conditions to simulate}

\item{ncpeaks}{percentage of peaks influenced by conditions}

\item{nbatch}{Number of batches to simulate}

\item{nbpeaks}{percentage of peaks influenced by batchs}

\item{npercond}{Number of samples per condition to simulate}

\item{nperbatch}{Number of samples per batch to simulate}

\item{seed}{Random seed for reproducibility}
}
\value{
list with rtmz data matrix, row index of peaks influenced by conditions, row index of peaks influenced by batchs, column index of conditions, column of batchs, raw condition matrix, raw batch matrix, peak mean across the samples, peak rsd across the samples
}
\description{
Simulation from data by sample mean and rsd from Empirical Cumulative Distribution or Bootstrap sampling
}
\details{
the numbers of batch columns should be the same with the condition columns.
}
\examples{
library(faahKO)
cdfpath <- system.file("cdf", package = "faahKO")
list <- getmr(cdfpath, pmethod = ' ')
sim <- simmzrt(list$data)
}
