% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{svacor}
\alias{svacor}
\title{Use Surrogate Variable Analysis(SVA) to correct the unknown batch effects}
\usage{
svacor(data, lv)
}
\arguments{
\item{data}{data as mzrt profile}

\item{lv}{factor vector for the group infomation}
}
\value{
list object with various components such raw data, corrected data, signal part, random errors part, batch part, p-values, q-values, mass, rt, Posterior Probabilities of Surrogate variables and Posterior Probabilities of Mod. If no surrogate variable found, corresponding part would miss.
}
\description{
Use Surrogate Variable Analysis(SVA) to correct the unknown batch effects
}
\details{
this is used for SVA to correct the unknown batch effects
}
\examples{
\dontrun{
library(faahKO)
cdfpath <- system.file("cdf", package = "faahKO")
cdffiles <- list.files(cdfpath, recursive = TRUE, full.names = TRUE)
list <- getmr(cdfpath, pmethod = ' ')
list <- svacor(list$data,list$group$class)
}
}
\seealso{
\code{\link{svapca}}, \code{\link{svaplot}}, \code{\link{svabatch}}
}
