% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sva.R
\name{svadata}
\alias{svadata}
\title{Filter the data with p value and q value}
\usage{
svadata(list, pqvalues = "sv", pt = 0.05, qt = 0.05)
}
\arguments{
\item{list}{results from svacor function}

\item{pqvalues}{method for ANOVA or SVA}

\item{pt}{threshold for p value, default is 0.05}

\item{qt}{threshold for q value, default is 0.05}
}
\value{
data, corrected data, mz and retention for fileted data
}
\description{
Filter the data with p value and q value
}
