% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sva.R
\name{svapca}
\alias{svapca}
\title{Principal component analysis(PCA) for SVA corrected data and raw data}
\usage{
svapca(list, center = T, scale = T, lv = NULL)
}
\arguments{
\item{list}{results from svacor function}

\item{center}{parameters for PCA}

\item{scale}{parameters for scale}

\item{lv}{group information}
}
\value{
plot
}
\description{
Principal component analysis(PCA) for SVA corrected data and raw data
}
