% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sva.R
\name{svaplot}
\alias{svaplot}
\title{Filter the data with p value and q value and show them}
\usage{
svaplot(list, pqvalues = "sv", pt = 0.05, qt = 0.05, lv = NULL,
  index = NULL)
}
\arguments{
\item{list}{results from svacor function}

\item{pqvalues}{method for ANOVA or SVA}

\item{pt}{threshold for p value, default is 0.05}

\item{qt}{threshold for q value, default is 0.05}

\item{lv}{group information}

\item{index}{index for selected peaks}
}
\value{
heatmap for the data
}
\description{
Filter the data with p value and q value and show them
}
