% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sva.R
\name{svaupload}
\alias{svaupload}
\title{Get the corrected data after SVA for metabolanalyst}
\usage{
svaupload(xset, lv = NULL)
}
\arguments{
\item{xset}{xcmsset object}

\item{lv}{group information}
}
\value{
csv files for both raw and corrected data for metabolanalyst if SVA could be applied
}
\description{
Get the corrected data after SVA for metabolanalyst
}
