% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda.R
\docType{data}
\name{TBBPA}
\alias{TBBPA}
\title{Demo data for TBBPA metabolism in Pumpkin}
\format{A list object with data, mass to charge ratio, retention time and group information. Three pumpkin seeding root samples' peaks list is extracted by xcms online.}
\usage{
data(TBBPA)
}
\description{
Demo data for TBBPA metabolism in Pumpkin
}
\references{
Hou, X., Yu, M., Liu, A., Wang, X., Li, Y., Liu, J., Schnoor, J.L., Jiang, G., 2019. Glycosylation of Tetrabromobisphenol A in Pumpkin. Environ. Sci. Technol. https://doi.org/10.1021/acs.est.9b02122
}
\keyword{datasets}
