% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{xrankanno}
\alias{xrankanno}
\title{Perform MS/MS X rank annotation for mgf file}
\usage{
xrankanno(file, db = NULL, ppm = 10, prems = 1.1, intc = 0.1, quantile = 0.75)
}
\arguments{
\item{file}{mgf file generated from MS/MS data}

\item{db}{database could be list object from `getms2pmd`}

\item{ppm}{mass accuracy, default 10}

\item{prems}{precursor mass range, default 1.1 to include M+H or M-H}

\item{intc}{intensity cutoff for peaks. Default 0.1}

\item{quantile}{X rank quantiles cutoff for annotation. Default 0.75.}
}
\value{
list with MSMS annotation results
}
\description{
Perform MS/MS X rank annotation for mgf file
}
