% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{findline}
\alias{findline}
\title{find line of the regression model for GC-MS}
\usage{
findline(data, threshold = 2, temp = c(100, 320))
}
\arguments{
\item{data}{imported data matrix of GC-MS}

\item{threshold}{the threshold of the response (log based 10)}

\item{temp}{the scale of the oven temperature (constant rate)}
}
\value{
list linear regression model for the matrix
}
\description{
find line of the regression model for GC-MS
}
\examples{
\dontrun{
data(matrix)
findline(matrix)
}
}
