% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{getfeaturesanova}
\alias{getfeaturesanova}
\title{Get the features from anova, with p value, q value, rsd and power restriction}
\usage{
getfeaturesanova(list, power = 0.8, pt = 0.05, qt = 0.05, n = 3,
  ng = 3, rsdcf = 100, inscf = 5, imputation = "l", index = NULL)
}
\arguments{
\item{list}{list with data as peaks list, mz, rt and group information (more than two groups)}

\item{power}{defined power}

\item{pt}{p value threshold}

\item{qt}{q value threshold, BH adjust}

\item{n}{sample numbers in one group}

\item{ng}{group numbers}

\item{rsdcf}{the rsd cutoff of all peaks in all group}

\item{inscf}{Log intensity cutoff for peaks across samples. If any peaks show a intensity higher than the cutoff in any samples, this peaks would not be filtered. default 5}

\item{imputation}{parameters for `getimputation` function method}

\item{index}{the index of peaks considered, default NULL}
}
\value{
dataframe with peaks fit the setting above
}
\description{
Get the features from anova, with p value, q value, rsd and power restriction
}
