% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcmsplus.R
\name{getfeaturesanova}
\alias{getfeaturesanova}
\title{Get the features from anova, with p value, q value, rsd and power restriction}
\usage{
getfeaturesanova(xod, power = 0.8, pt = 0.05, qt = 0.05, n = 3,
  ng = 3, rsdt = 30)
}
\arguments{
\item{xod}{xcmsSet objects}

\item{power}{defined power}

\item{pt}{p value threshold}

\item{qt}{q value threshold, BH adjust}

\item{n}{sample numbers in one group}

\item{ng}{group numbers}

\item{rsdt}{rsd threshold to filter the data}
}
\value{
dataframe with peaks fit the setting above
}
\description{
Get the features from anova, with p value, q value, rsd and power restriction
}
