% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{getfeaturest}
\alias{getfeaturest}
\title{Get the features from t test, with p value, q value, rsd and power restriction}
\usage{
getfeaturest(list, power = 0.8, pt = 0.05, qt = 0.05, n = 3, imputation = "l")
}
\arguments{
\item{list}{list with data as peaks list, mz, rt and group information (two groups)}

\item{power}{defined power}

\item{pt}{p value threshold}

\item{qt}{q value threshold, BH adjust}

\item{n}{sample numbers in one group}

\item{imputation}{parameters for `getimputation` function method}
}
\value{
dataframe with peaks fit the setting above
}
\description{
Get the features from t test, with p value, q value, rsd and power restriction
}
