% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{getformula}
\alias{getformula}
\title{Get chemical formula for mass to charge ratio.}
\usage{
getformula(
  mz,
  charge = 0,
  window = 0.001,
  elements = list(C = c(1, 50), H = c(1, 50), N = c(0, 50), O = c(0, 50), P = c(0, 1),
    S = c(0, 1))
)
}
\arguments{
\item{mz}{a vector with mass to charge ratio}

\item{charge}{The charge value of the formula, default 0 for autodetect}

\item{window}{The window accuracy in the same units as mass}

\item{elements}{Elements list to take into account.}
}
\value{
list with chemical formula
}
\description{
Get chemical formula for mass to charge ratio.
}
