% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzrt.R
\name{getimputation}
\alias{getimputation}
\title{Impute the peaks list data}
\usage{
getimputation(list, method = "l")
}
\arguments{
\item{list}{list with data as peaks list, mz, rt and group information}

\item{method}{'r' means remove, 'l' means use half the minimum of the values across the peaks list, 'mean' means mean of the values across the samples, 'median' means median of the values across the samples, '0' means 0, '1' means 1. Default 'l'.}
}
\value{
list with imputed peaks
}
\description{
Impute the peaks list data
}
\examples{
\dontrun{
library(faahKO)
cdfpath <- system.file('cdf', package = 'faahKO')
list <- getmr(cdfpath, pmethod = ' ')
getimputation(list)
}
}
\seealso{
\code{\link{getdata2}},\code{\link{getdata}}, \code{\link{getmzrt}},\code{\link{getmzrt2}}, \code{\link{getdoe}}, \code{\link{getmr}}
}
