% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.R
\name{GetIntegration}
\alias{GetIntegration}
\title{GetIntegration was mainly used for get the intergration of certain ion's chromatogram data and plot the data}
\usage{
GetIntegration(data, minrt = 8.3, maxrt = 9, n = 5, m = 5,
  startslope = 2, stopslope = 2, baseline = 20, noslope = T,
  smoothit = T, name = NULL, half = F, ...)
}
\arguments{
\item{data}{file should be a dataframe with the first column RT and second column intensity of the SIM ions.}

\item{minrt}{a rough smallest RT range contained only one peak to get the area}

\item{maxrt}{a rough largest RT range contained only one peak to get the area}

\item{n}{points in the moving average smooth box, default value is 5}

\item{m}{numbers of points for regression to get the slope}

\item{startslope}{the threshold value for start peak (%max slope)}

\item{stopslope}{the threshold value for stop peak (%max slope)}

\item{baseline}{numbers of the points for the baseline of the signal}

\item{noslope}{logical, if using a horizon line to get area or not}

\item{smoothit}{logical, if using an average smooth box or not. If using, n will be used}

\item{half}{logical, if using the left half peak to caculate the area}
}
\value{
list intergration data such as peak area, peak hight, signal and the slope data.
}
\description{
GetIntegration was mainly used for get the intergration of certain ion's chromatogram data and plot the data
}
\examples{
NULL
}

