% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.R
\name{Getisotopologues}
\alias{Getisotopologues}
\title{Get the selected isotopologues at certain MS data}
\usage{
Getisotopologues(formula = "C12OH6Br4", charge = "1", width = 0.3)
}
\description{
Get the selected isotopologues at certain MS data
}

