% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.R
\name{getmassdefect}
\alias{getmassdefect}
\title{Get mass defect with certain scaled factor}
\usage{
getmassdefect(mass, sf)
}
\arguments{
\item{mass}{vector of mass}

\item{sf}{scaled factors}
}
\value{
dataframe with mass, scaled mass and scaled mass defect
}
\description{
Get mass defect with certain scaled factor
}
\examples{
mass <- c(100.1022,245.2122,267.3144,400.1222,707.2294)
sf <- 0.9988
mf <- getmassdefect(mass,sf)
}
