% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcmsplus.R
\name{getmassdiff}
\alias{getmassdiff}
\title{Isotope extraction for single group of samples with certain mass diff}
\usage{
getmassdiff(xcmsSet, massdiff, rtwindow, mzwindow, ppm)
}
\arguments{
\item{xcmsSet}{a xcmsSet object}

\item{massdiff}{mass defection}

\item{rtwindow}{retention time range}

\item{mzwindow}{mass charge ratio window}

\item{ppm}{resolution of the mass spectrum}
}
\value{
table with mass, retention time, scaled mass and scaled mass defect
}
\description{
Isotope extraction for single group of samples with certain mass diff
}
