% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{getmd}
\alias{getmd}
\title{import data and return the annotated matrix for GC-MS}
\usage{
getmd(data, profstep = 1, time = 1)
}
\arguments{
\item{data}{file type which xcmsRaw could handle}

\item{profstep}{the m/z step for generating matrix data from raw mass spectral data.}

\item{time}{round numbers of retention time, default is 1}
}
\value{
matrix with the row as increasing m/z second and column as increasing scantime
}
\description{
import data and return the annotated matrix for GC-MS
}

