% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{getmzrt}
\alias{getmzrt}
\title{Get the mzrt profile and group information for batch correction and plot as a list}
\usage{
getmzrt(xset, name = NULL, mzdigit = 4, rtdigit = 1, eic = F)
}
\arguments{
\item{xset}{xcmsSet/XCMSnExp objects}

\item{name}{file name for csv file, default NULL}

\item{mzdigit}{m/z digits of row names of data frame}

\item{rtdigit}{retention time digits of row names of data frame}

\item{eic}{logical, save xcmsEIC objects for further investigation with the same name of files, you will need raw files in the same directory as defined in xcmsSet to extract the EIC based on the binned data. default F}
}
\value{
list with rtmz profile and group infomation
}
\description{
Get the mzrt profile and group information for batch correction and plot as a list
}
\examples{
\dontrun{
library(faahKO)
cdfpath <- system.file('cdf', package = 'faahKO')
xset <- getdata(cdfpath, pmethod = ' ')
getmzrt(xset)
}
}
\seealso{
\code{\link{getdata}},\code{\link{getupload}}, \code{\link{getdoe}},\code{\link{getmzrt}}
}
