% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{getmzrtcsv}
\alias{getmzrtcsv}
\title{Covert the peaks list csv file into list}
\usage{
getmzrtcsv(path)
}
\arguments{
\item{path}{the path to your csv file}
}
\value{
list with rtmz profile and group infomation
}
\description{
Covert the peaks list csv file into list
}
\seealso{
\code{\link{getmzrt}}, \code{\link{getmzrt2}}
}
