% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmzrt.R
\name{getoverlaprt}
\alias{getoverlaprt}
\title{Get the overlap peaks by retention time}
\usage{
getoverlaprt(rtrange1, rtrange2)
}
\arguments{
\item{rtrange1}{mass range 1 to be overlapped}

\item{rtrange2}{mass range 2 to overlap}
}
\value{
logical index for rtrange1's peaks
}
\description{
Get the overlap peaks by retention time
}
\seealso{
\code{\link{getmzrt}}, \code{\link{getimputation}}, \code{\link{getmr}},\code{\link{getdoe}}, \code{\link{getoverlapmass}},\code{\link{getoverlappeak}}
}
