% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalstd.R
\name{getpaired}
\alias{getpaired}
\title{Filter ions/peaks based on retention time hierarchical clustering, paired mass differences(PMD) and PMD frequency analysis.}
\usage{
getpaired(list, rtcutoff = 9, freqcutoff = 30, pmdcutoff = 100)
}
\arguments{
\item{list}{a list with mzrt profile}

\item{rtcutoff}{cutoff of the distances in retention time hierarchical clustering analysis, default 9}

\item{freqcutoff}{cutoff of the paired mass difference frequency, default 30}

\item{pmdcutoff}{cutoff of the largest paired mass differences, default 100}
}
\value{
list with tentative isotope, adducts, and neutral loss peaks' index, retention time clusters.
}
\description{
Filter ions/peaks based on retention time hierarchical clustering, paired mass differences(PMD) and PMD frequency analysis.
}
\seealso{
\code{\link{getstd}},\code{\link{getsda}},\code{\link{plotpaired}}
}
