% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massdefect.R
\name{getpaired}
\alias{getpaired}
\title{Paired mass diff relationship among peak list based on cluster analysis}
\usage{
getpaired(list, rtcutoff = 9, isocutoff = 3, freqcutoff = 20,
  submass = c(15.9949, 14.003074, 26.01568, 14.01565, 43.00581, 30.01056,
  34.96885, 78.91834), mdcutoff = 0.02)
}
\arguments{
\item{list}{a list with mzrt profile}

\item{rtcutoff}{cutoff of the distances in cluster}

\item{isocutoff}{cutoff to find the isotope relationship}

\item{freqcutoff}{cutoff of the mass differences frequency}

\item{submass}{mass vector of sub structure of homologous series}

\item{mdcutoff}{mass defect cluster cutoff}
}
\value{
list with tentative isotope, adducts, and neutral loss peaks' index, retention time cluster, std mass defect analysis dataframe
}
\description{
Paired mass diff relationship among peak list based on cluster analysis
}
\seealso{
\code{\link{getmassdefect}},\code{\link{plotkms}},\code{\link{getcorstd}}
}
