% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massdefect.R
\name{getpaired}
\alias{getpaired}
\title{Filter ions/peaks based on retention time hierarchical clustering, paired mass differences(PMD) and PMD frequency analysis.}
\usage{
getpaired(list, rtcutoff = 9, freqcutoff = 30, pmdcutoff = 100)
}
\arguments{
\item{list}{a list with mzrt profile}

\item{rtcutoff}{cutoff of the distances in cluster, default 9}

\item{freqcutoff}{cutoff of the mass differences frequency, default 30}

\item{pmdcutoff}{cutoff of the largest mass differences, default 100}
}
\value{
list with tentative isotope, adducts, and neutral loss peaks' index, retention time clusters.
}
\description{
Filter ions/peaks based on retention time hierarchical clustering, paired mass differences(PMD) and PMD frequency analysis.
}
\seealso{
\code{\link{getmassdefect}},\code{\link{getstd}},\code{\link{getstd}},\code{\link{getmdg}},\code{\link{plotpaired}}
}
