% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmzrt.R
\name{getpower}
\alias{getpower}
\title{Get the index with power restriction for certain study with BH adjusted p-value and certain power.}
\usage{
getpower(list, pt = 0.05, qt = 0.05, powert = 0.8,
  imputation = "l")
}
\arguments{
\item{list}{list with data as peaks list, mz, rt and group information}

\item{pt}{p value threshold, default 0.05}

\item{qt}{q value threshold, BH adjust, default 0.05}

\item{powert}{power cutoff, default 0.8}

\item{imputation}{parameters for `getimputation` function method}
}
\value{
list with current power and sample numbers for each peaks
}
\description{
Get the index with power restriction for certain study with BH adjusted p-value and certain power.
}
\examples{
data(list)
getpower(list)
}
\seealso{
\code{\link{getdata2}},\code{\link{getdata}}, \code{\link{getmzrt}}, \code{\link{getimputation}}, \code{\link{getmr}},\code{\link{getdoe}}
}
