% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmzrt.R
\name{getretcor}
\alias{getretcor}
\title{Perform peaks list alignment and return features table}
\usage{
getretcor(list, ts = 1, ppm = 10, deltart = 5, FUN)
}
\arguments{
\item{list}{each element should be a data.frame with mz, rt and ins as m/z, retention time in seconds and intensity of certain peaks.}

\item{ts}{template sample index in the list, default 1}

\item{ppm}{mass accuracy, default 10}

\item{deltart}{retention time shift table, default 5 seconds}

\item{FUN}{function to deal with multiple aligned peaks from one sample}
}
\value{
mzrt object without group information
}
\description{
Perform peaks list alignment and return features table
}
