% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcmsdoe.R
\name{gettechrep}
\alias{gettechrep}
\title{Get the report for technique replicates.}
\usage{
gettechrep(xset, method = "medret", intensity = "into", file = NULL,
  rsdcf = 30, inscf = 5)
}
\arguments{
\item{xset}{the xcmsset object which for all of your technique replicates for one sample}

\item{method}{parameter for groupval function}

\item{intensity}{parameter for groupval function}

\item{file}{file name for further annotation, default NULL}

\item{rsdcf}{rsd cutoff for peaks, default 30}

\item{inscf}{Log intensity cutoff for peaks, default 5}
}
\value{
dataframe with mean, standard deviation and RSD for those technique replicates combined with raw data
}
\description{
Get the report for technique replicates.
}
\examples{
\dontrun{
data(fish4)
fish <- as(fish4,"xcmsSet")
gettechrep(fish)
}
}
