% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{gettechrep}
\alias{gettechrep}
\title{Get the report for technique replicates.}
\usage{
gettechrep(xset, method = "medret", intensity = "into", file = NULL,
  rsdcf = 30, inscf = 1000)
}
\arguments{
\item{xset}{the xcmsset object which for all of your technique replicates for one sample}

\item{method}{parameter for groupval function}

\item{intensity}{parameter for groupval function}

\item{file}{file name for further annotation, default NULL}

\item{rsdcf}{rsd cutoff for peaks, default 30}

\item{inscf}{intensity cutoff for peaks, default 1000}
}
\value{
dataframe with mean, standard deviation and RSD for those technique replicates combined with raw data
}
\description{
Get the report for technique replicates.
}
