% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{getupload}
\alias{getupload}
\title{Get the csv files from xcmsset/XCMSnExp/list object}
\usage{
getupload(xset, method = "medret", value = "into", name = "Peaklist",
  type = "m", mzdigit = 4, rtdigit = 1)
}
\arguments{
\item{xset}{the xcmsset/XCMSnExp/list object which you want to submitted to Metaboanalyst}

\item{method}{parameter for groupval function}

\item{value}{parameter for groupval function}

\item{name}{file name}

\item{type}{m means  Metaboanalyst, a means xMSannotator, o means full infomation csv}

\item{mzdigit}{m/z digits of row names of data frame}

\item{rtdigit}{retention time digits of row names of data frame}
}
\value{
dataframe with data needed for Metaboanalyst/xMSannotator/pmd if your want to perform local analysis.
}
\description{
Get the csv files from xcmsset/XCMSnExp/list object
}
\examples{
\dontrun{
library(faahKO)
cdfpath <- system.file('cdf', package = 'faahKO')
xset <- getdata(cdfpath, pmethod = ' ')
getupload(xset)
}
}
\seealso{
\code{\link{getdata}}, \code{\link{getmzrt}}
}
