% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{getupload}
\alias{getupload}
\title{Get the csv files to be submitted to Metaboanalyst}
\usage{
getupload(xset, method = "medret", value = "into", name = "Peaklist")
}
\arguments{
\item{xset}{the xcmsset object which you want to submitted to Metaboanalyst}

\item{method}{parameter for groupval function}

\item{value}{parameter for groupval function}

\item{name}{file name}
}
\value{
dataframe with data needed for Metaboanalyst if your want to perform local analysis.
}
\description{
Get the csv files to be submitted to Metaboanalyst
}
\examples{
\dontrun{
library(faahKO)
cdfpath <- system.file('cdf', package = 'faahKO')
xset <- getdata(cdfpath, pmethod = ' ')
getupload(xset)
}
}
\seealso{
\code{\link{getdata}},\code{\link{getupload2}}, \code{\link{getmzrt}}
}
