% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcmsplot.R
\name{gifmr}
\alias{gifmr}
\title{plot scatter plot for rt-mz profile and output gif file for mutiple groups}
\usage{
gifmr(xset, rsdcf = 30, inscf = 5, file = "test")
}
\arguments{
\item{xset}{xcmsset object with mutiple groups}

\item{rsdcf}{the rsd cutoff of all peaks}

\item{inscf}{log intensity cutoff for peaks, default 5}

\item{file}{file name for further annotation, default NULL}

\item{...}{parameters for `plot` function}
}
\value{
gif file and csv file
}
\description{
plot scatter plot for rt-mz profile and output gif file for mutiple groups
}
\examples{
\dontrun{
library(faahKO)
cdfpath <- system.file("cdf", package = "faahKO")
xset <- getdata(cdfpath, pmethod = ' ')
gifmr(xset)
}
}
