% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massdefect.R
\name{globalstd}
\alias{globalstd}
\title{GlobalStd algorithm}
\usage{
globalstd(list, rtcutoff = 9, freqcutoff = 20, corcutoff = NULL,
  submass = c(15.9949, 14.003074, 26.01568, 14.01565, 43.00581, 30.01056,
  34.96885, 78.91834), mdgn = 20, lv = NULL)
}
\arguments{
\item{list}{a peaks list with mass to charge}

\item{rtcutoff}{cutoff of the distances in cluster}

\item{freqcutoff}{cutoff of the mass differences frequency}

\item{corcutoff}{cutoff of the correlation coefficient, default NULL}

\item{submass}{mass vector of sub structure of homologous series}

\item{mdgn}{mass defect groups numbers for interval, 20 means 0.05 inteval on mass defect scale from -0.5 to 0.5}

\item{lv}{group info for the data}
}
\value{
list with GlobalStd algorithm processed data.
}
\description{
GlobalStd algorithm
}
\seealso{
\code{\link{getpaired}},\code{\link{getstd}},\code{\link{getmdg}},\code{\link{plotstd}},\code{\link{plotstdmd}},\code{\link{plotstdrt}}
}
