% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalstd.R
\name{globalstd}
\alias{globalstd}
\title{GlobalStd algorithm}
\usage{
globalstd(list, rtcutoff = 9, freqcutoff = 30, corcutoff = NULL,
  pmdcutoff = 100, rtcutoff2 = 3, freqcutoff2 = 10, pmdcutoff2 = 500)
}
\arguments{
\item{list}{a peaks list with mass to charge}

\item{rtcutoff}{cutoff of the distances in cluster}

\item{freqcutoff}{cutoff of the mass differences frequency}

\item{corcutoff}{cutoff of the correlation coefficient, default NULL}

\item{pmdcutoff}{cutoff of the largest mass differences, default 100}

\item{rtcutoff2}{cutoff of the retention time in secounds, default 3}

\item{freqcutoff2}{cutoff of the paired mass difference frequency, default 10}

\item{pmdcutoff2}{cutoff of the largest paired mass differences, default 500}
}
\value{
list with GlobalStd algorithm processed data.
}
\description{
GlobalStd algorithm
}
\seealso{
\code{\link{getpaired}},\code{\link{getstd}},\code{\link{getsda}},\code{\link{plotstd}},\code{\link{plotstdsda}},\code{\link{plotstdrt}}
}
