% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzrtsim.R
\name{mzrtsim}
\alias{mzrtsim}
\title{Generate simulated count data with batch effects for npeaks}
\usage{
mzrtsim(npeaks = 1000, ncond = 2, ncpeaks = 0.05, nbatch = 3,
  nbpeaks = 0.1, npercond = 10, nperbatch = c(8, 5, 7), shape = 8,
  scale = 12, seed = 42)
}
\arguments{
\item{npeaks}{Number of peaks to simulate}

\item{ncond}{Number of conditions to simulate}

\item{ncpeaks}{percentage of peaks influenced by conditions}

\item{nbatch}{Number of batches to simulate}

\item{nbpeaks}{percentage of peaks influenced by batchs}

\item{npercond}{Number of samples per condition to simulate}

\item{nperbatch}{Number of samples per batch to simulate}

\item{shape}{shape for Weibull distribution}

\item{scale}{scale for Weibull distribution}

\item{seed}{Random seed for reproducibility}
}
\value{
list with rtmz data matrix, row index of peaks influenced by conditions, row index of peaks influenced by batchs, column index of conditions, column of batchs, raw condition matrix, raw batch matrix
}
\description{
Generate simulated count data with batch effects for npeaks
}
\details{
the numbers of batch columns should be the same with the condition columns.
}
\examples{
sim <- mzrtsim()
}
