% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmzrt.R
\name{plotden}
\alias{plotden}
\title{plot the density for multiple samples}
\usage{
plotden(data, lv, index = NULL, name = NULL, ...)
}
\arguments{
\item{data}{mzrt profile with row peaks and column samples}

\item{lv}{group information}

\item{index}{index for selected peaks}

\item{name}{name on the figure for samples}

\item{...}{parameters for `plot` function}
}
\description{
plot the density for multiple samples
}
\examples{
data(list)
plotden(list$data, lv = as.character(list$group$class))
}
