% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{plothist}
\alias{plothist}
\title{plot the density of the GC-MS data with EM algorithm to seperate the data into two log normal distribution.}
\usage{
plothist(data)
}
\arguments{
\item{data}{imported data matrix of GC-MS}
}
\description{
plot the density of the GC-MS data with EM algorithm to seperate the data into two log normal distribution.
}

