% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{plotint}
\alias{plotint}
\title{plot the information of intergretion}
\usage{
plotint(list, name = NULL)
}
\arguments{
\item{list}{list from getinteragtion}

\item{name}{the title of the plot}
}
\description{
plot the information of intergretion
}

