% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{plotkms}
\alias{plotkms}
\title{plot the kendrick mass defect diagram}
\usage{
plotkms(data, cutoff = 1000)
}
\arguments{
\item{data}{vector with the name m/z}

\item{cutoff}{remove the low intensity}
}
\description{
plot the kendrick mass defect diagram
}
\examples{
\dontrun{
mz <- c(10000,5000,20000,100,40000)
names(mz) <- c(100.1022,245.2122,267.3144,400.1222,707.2294)
plotkms(mz)
}
}
\seealso{
\code{\link{getmassdefect}}
}
