% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcmsplot.R
\name{plotmr}
\alias{plotmr}
\title{plot the scatter plot for xcmsset objects with threshold}
\usage{
plotmr(xset, ms = c(100, 800), inscf = 5, rsdcf = 30, ...)
}
\arguments{
\item{xset}{the xcmsset object}

\item{ms}{the mass range to plot the data}

\item{inscf}{Log intensity cutoff for peaks, default 5}

\item{rsdcf}{the rsd cutoff of all peaks}

\item{...}{parameters for `plot` function}
}
\value{
data fit the cutoff
}
\description{
plot the scatter plot for xcmsset objects with threshold
}
\examples{
\dontrun{
library(faahKO)
cdfpath <- system.file("cdf", package = "faahKO")
xset <- getdata(cdfpath, pmethod = ' ')
plotmr(xset)
}
}
