% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcmsplus.R
\name{plotmr}
\alias{plotmr}
\title{plot the scatter plot for xcmsset (or two) objects with threshold}
\usage{
plotmr(data1, data2 = NULL, threshold = 5, ms = c(100, 1000), ...)
}
\arguments{
\item{data1}{the first xcmsset object}

\item{data2}{the second xcmsset object}

\item{threshold}{the threshold of the response (log based 10)}

\item{ms}{the mass range to plot the data}

\item{...}{parameters for `plot` function}
}
\description{
plot the scatter plot for xcmsset (or two) objects with threshold
}
\examples{
\dontrun{
library(faahKO)
cdfpath <- system.file("cdf", package = "faahKO")
xset <- getdata(cdfpath, pmethod = ' ')
plotmr(xset)
}
}
