% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcmsplus.R
\name{plotmrc}
\alias{plotmrc}
\title{plot the diff scatter plot for one xcmsset objects with threshold and two groups}
\usage{
plotmrc(data, threshold = 5, ms = c(100, 1000), ...)
}
\arguments{
\item{data}{xcmsset object with two groups}

\item{threshold}{the threshold of the response (log based 10)}

\item{ms}{the mass range to plot the data}

\item{...}{parameters for `plot` function}
}
\description{
plot the diff scatter plot for one xcmsset objects with threshold and two groups
}
\examples{
\dontrun{
library(faahKO)
cdfpath <- system.file("cdf", package = "faahKO")
xset <- getdata(cdfpath, pmethod = ' ')
plotmrc(xset)
}
}
