% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzrt.R
\name{plotmrc}
\alias{plotmrc}
\title{plot the diff scatter plot for one xcmsset objects with threshold between two groups}
\usage{
plotmrc(list, ms = c(100, 800), inscf = 5, rsdcf = 30, imputation = "l",
  index = NULL, ...)
}
\arguments{
\item{list}{list with data as peaks list, mz, rt and group information}

\item{ms}{the mass range to plot the data}

\item{inscf}{Log intensity cutoff for peaks across samples. If any peaks show a intensity higher than the cutoff in any samples, this peaks would not be filtered. default 5}

\item{rsdcf}{the rsd cutoff of all peaks in all group}

\item{imputation}{parameters for `getimputation` function method}

\item{index}{the index of peaks considered, default NULL}

\item{...}{parameters for `plot` function}
}
\description{
plot the diff scatter plot for one xcmsset objects with threshold between two groups
}
\examples{
\dontrun{
library(faahKO)
cdfpath <- system.file('cdf', package = 'faahKO')
list <- getmr(cdfpath, pmethod = ' ')
plotmrc(list)
}
}
