% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{plotms}
\alias{plotms}
\title{plot GC/LC-MS data as a heatmap with TIC}
\usage{
plotms(data, log = F)
}
\arguments{
\item{data}{imported data matrix of GC-MS}

\item{log}{transform the intensity into log based 10}
}
\value{
heatmap
}
\description{
plot GC/LC-MS data as a heatmap with TIC
}
\examples{
\dontrun{
library(faahKO)
cdfpath <- system.file('cdf', package = 'faahKO')
cdffiles <- list.files(cdfpath, recursive = TRUE, full.names = TRUE)
matrix <- getmd(cdffiles[1])
png('test.png')
plotms(matrix)
dev.off()
}
}
