% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{plotms}
\alias{plotms}
\title{plot GC/LC-MS data as a heatmap with TIC}
\usage{
plotms(data, log = FALSE)
}
\arguments{
\item{data}{imported data matrix of GC-MS}

\item{log}{transform the intensity into log based 10}
}
\value{
heatmap
}
\description{
plot GC/LC-MS data as a heatmap with TIC
}
\examples{
\dontrun{
png('test.png')
plotms(matrix)
dev.off()
}
}
