% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{plotmsrt}
\alias{plotmsrt}
\title{Plot EIC of certain m/z and return dataframe for integration}
\usage{
plotmsrt(data, ms, rt, n = FALSE)
}
\arguments{
\item{data}{imported data matrix of GC-MS}

\item{ms}{m/z to be extracted}

\item{rt}{vector range of the retention time}

\item{n}{logical smooth or not}
}
\value{
dataframe with  with the first column RT and second column intensity of the SIM ions.
}
\description{
Plot EIC of certain m/z and return dataframe for integration
}
\examples{
\dontrun{
matrix <- getmd(rawdata)
plotmsrt(matrix,rt = c(500,1000),ms = 300)
}
}
