% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{plotmz}
\alias{plotmz}
\title{plot GC/LC-MS data as scatter plot}
\usage{
plotmz(data, inscf = 3.5, ...)
}
\arguments{
\item{data}{imported data matrix of GC-MS}

\item{inscf}{Log intensity cutoff for peaks, default 3.5}

\item{...}{parameters for `plot` function}
}
\value{
scatter plot
}
\description{
plot GC/LC-MS data as scatter plot
}
\examples{
\dontrun{
data(matrix)
png('test.png')
plotmz(matrix)
dev.off()
}
}
