% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalstd.R
\name{plotpaired}
\alias{plotpaired}
\title{Plot the mass pairs and high frequency mass differences}
\usage{
plotpaired(list)
}
\arguments{
\item{list}{a list from getpaired function}
}
\description{
Plot the mass pairs and high frequency mass differences
}
\seealso{
\code{\link{getpaired}}, \code{\link{globalstd}}
}
