% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzrt.R
\name{plotpca}
\alias{plotpca}
\title{plot the PCA of list}
\usage{
plotpca(data, lv = NULL, index = NULL, center = T, scale = T, ...)
}
\arguments{
\item{data}{mzrt profile with row peaks and column samples}

\item{lv}{group information}

\item{index}{index for selected peaks}

\item{center}{parameters for PCA}

\item{scale}{parameters for scale}

\item{...}{other parameters for `plot` function}
}
\description{
plot the PCA of list
}
\examples{
\dontrun{
library(faahKO)
cdfpath <- system.file('cdf', package = 'faahKO')
list <- getmr(cdfpath, pmethod = ' ')
data <- list$data
lv <- as.character(list$group$class)
plotpca(data, lv)
}
}
