% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcmsplus.R
\name{plotpca}
\alias{plotpca}
\title{plot the PCA of xcmsset}
\usage{
plotpca(xset, lv = NULL, center = T, scale = T, ...)
}
\arguments{
\item{xset}{xcmsset data}

\item{lv}{group information}

\item{center}{parameters for PCA}

\item{scale}{parameters for scale}

\item{...}{other parameters for `plot` function}
}
\description{
plot the PCA of xcmsset
}
\examples{
\dontrun{
library(faahKO)
cdfpath <- system.file("cdf", package = "faahKO")
xset <- getdata(cdfpath, pmethod = ' ')
plotpca(xset)
}
}
