% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmzrt.R
\name{plotridge}
\alias{plotridge}
\title{plot ridgeline density plot}
\usage{
plotridge(data, lv = NULL, indexx = NULL, indexy = NULL, ...)
}
\arguments{
\item{data}{matrix}

\item{lv}{factor vector for the column}

\item{indexx}{index for matrix row}

\item{indexy}{index for matrix column}

\item{...}{parameters for `title` function}
}
\value{
ridgeline density plot
}
\description{
plot ridgeline density plot
}
\examples{
data(list)
plotridge(t(list$data),indexy=c(1:10),xlab = 'Intensity',ylab = 'peaks')
plotridge(log(list$data),as.factor(list$group$sample_group),xlab = 'Intensity',ylab = 'peaks')
}
