% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcmsplus.R
\name{plotrsd}
\alias{plotrsd}
\title{plot the rsd influnces of data}
\usage{
plotrsd(xset, ...)
}
\arguments{
\item{xset}{xcmsset data}

\item{...}{other parameters for `plot` function}
}
\description{
plot the rsd influnces of data
}
\examples{
\dontrun{
library(faahKO)
cdfpath <- system.file("cdf", package = "faahKO")
xset <- getdata(cdfpath, pmethod = ' ')
plotrsd(xset)
}
}
