% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{plott}
\alias{plott}
\title{plot GC-MS data as a heatmap for constant speed of temperature rising}
\usage{
plott(data, log = FALSE, temp = c(100, 320))
}
\arguments{
\item{data}{imported data matrix of GC-MS}

\item{log}{transform the intensity into log based 10}

\item{temp}{temperature range for constant speed}
}
\value{
heatmap
}
\description{
plot GC-MS data as a heatmap for constant speed of temperature rising
}
\examples{
\dontrun{
matrix <- getmd(rawdata)
plott(matrix)
}
}
