% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{qbatch}
\alias{qbatch}
\title{Get the MIR from the file}
\usage{
qbatch(file, mz1, mz2, rt = c(8.65, 8.74), brt = c(8.74, 8.85))
}
\arguments{
\item{file}{data file, CDF or other format supportted by xcmsRaw}

\item{mz1}{the lowest mass}

\item{mz2}{the highest mass}

\item{rt}{a rough RT range contained only one peak to get the area}

\item{brt}{a rough RT range contained only one peak and enough noises to get the area}
}
\value{
arearatio
}
\description{
Get the MIR from the file
}
\examples{
\dontrun{
arearatio <- qbatch(datafile)
}
}
