% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{svapca}
\alias{svapca}
\title{Principal component analysis(PCA) for SVA/ISVA corrected data and raw data}
\usage{
svapca(list, center = T, scale = T, lv = NULL)
}
\arguments{
\item{list}{results from svacor function}

\item{center}{parameters for PCA}

\item{scale}{parameters for scale}

\item{lv}{factor vector for the group infomation}
}
\value{
plot
}
\description{
Principal component analysis(PCA) for SVA/ISVA corrected data and raw data
}
\examples{
\dontrun{
library(faahKO)
cdfpath <- system.file("cdf", package = "faahKO")
list <- getmr(cdfpath, pmethod = ' ')
li <- svacor(list$data,list$group$class)
svapca(li)
}
}
\seealso{
\code{\link{svacor}}, \code{\link{svaplot}}, \code{\link{svabatch}}
}
