% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{svaplot}
\alias{svaplot}
\title{Filter the data with p value and q value and show them}
\usage{
svaplot(list, lv, pqvalues = "sv", pt = 0.05, qt = 0.05, index = NULL)
}
\arguments{
\item{list}{results from svacor function}

\item{lv}{factor vector for the group infomation}

\item{pqvalues}{method for ANOVA or SVA}

\item{pt}{threshold for p value, default is 0.05}

\item{qt}{threshold for q value, default is 0.05}

\item{index}{index for selected peaks}
}
\value{
heatmap for the data
}
\description{
Filter the data with p value and q value and show them
}
\examples{
\dontrun{
library(faahKO)
cdfpath <- system.file("cdf", package = "faahKO")
list <- getmr(cdfpath, pmethod = ' ')
li <- svacor(list$data,list$group$class)
svaplot(li)
}
}
\seealso{
\code{\link{svacor}}, \code{\link{svapca}}, \code{\link{svabatch}}
}
