% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sva.R
\name{svaplot}
\alias{svaplot}
\title{Filter the data with p value and q value and show them}
\usage{
svaplot(list, pqvalues = "sv", pt = 0.05, qt = 0.05, lv = NULL,
  index = NULL)
}
\arguments{
\item{list}{results from svacor function}

\item{pqvalues}{method for ANOVA or SVA}

\item{pt}{threshold for p value, default is 0.05}

\item{qt}{threshold for q value, default is 0.05}

\item{lv}{group information}

\item{index}{index for selected peaks}
}
\value{
heatmap for the data
}
\description{
Filter the data with p value and q value and show them
}
\examples{
\dontrun{
library(faahKO)
cdfpath <- system.file("cdf", package = "faahKO")
cdffiles <- list.files(cdfpath, recursive = TRUE, full.names = TRUE)
xset <- xcmsSet(cdffiles)
xset <- group(xset)
xset2 <- retcor(xset, family = "symmetric", plottype = "mdevden")
xset2 <- group(xset2, bw = 10)
xset3 <- fillPeaks(xset2)
df <- svacor(xset3)
svaplot(df)
}
}
\seealso{
\code{\link{svacor}}, \code{\link{svapca}}, \code{\link{svabatch}}
}
