% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{writeMSP}
\alias{writeMSP}
\title{Write MSP files for NIST search}
\usage{
writeMSP(mz, outfilename = "unknown")
}
\arguments{
\item{mz}{a intensity vector, who name is the mass in m/z}

\item{outfilename}{the name of the MSP file, default is 'unknown'}
}
\value{
none a MSP file will be created at the subfolder working dictionary with name 'MSP'
}
\description{
Write MSP files for NIST search
}
\examples{
\dontrun{
mz <- c(10000,20000,10000,30000,5000)
names(mz) <- c(101,143,189,221,234)
writeMSP(mz,'test')
}
}
